backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
drf-yasg>=1.21,<2.0
Pillow>=10.0
python-decouple>=3.8,<4.0
PyMySQL>=1.1,<2.0
